rename metadrive type to type
