+++ metadrive/component/buildings/tollgate_building.py
@@ -1,5 +1,5 @@
 from metadrive.component.buildings.base_building import BaseBuilding
-from metadrive.metadrive_type import MetaDriveType
+from metadrive.type import MetaDriveType
 from metadrive.engine.asset_loader import AssetLoader
 from metadrive.utils.pg_utils.utils import generate_static_box_physics_body
