+++ metadrive/component/traffic_light/base_traffic_light.py
@@ -1,6 +1,6 @@
 from metadrive.base_class.base_object import BaseObject
 from metadrive.constants import MetaDriveType
-from metadrive.metadrive_type import TrafficLightStatus
+from metadrive.type import TrafficLightStatus
 from metadrive.engine.asset_loader import AssetLoader
 from metadrive.utils.pg_utils.utils import generate_static_box_physics_body
