+++ metadrive/manager/waymo_idm_traffic_manager.py
@@ -3,7 +3,7 @@ from metadrive.utils.waymo_utils.parse_object_state import parse_full_trajectory
 from collections import namedtuple, OrderedDict
 
 import numpy as np
-from metadrive.metadrive_type import MetaDriveType
+from metadrive.type import MetaDriveType
 from metadrive.component.lane.waymo_lane import PointLane
 from metadrive.component.vehicle.vehicle_type import SVehicle
 from metadrive.manager.waymo_traffic_manager import WaymoTrafficManager
